# backend/docs/openapi.apifox.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Light OSS API",
    "version": "1.0.0",
    "description": "OpenAPI 3.0 JSON for APIFOX import. Route coverage is derived from backend/internal/handler/router.go and excludes site content routes outside /api/v1 and /healthz."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local backend default"
    },
    {
      "url": "http://oss.underhear.cn",
      "description": "Gateway example"
    }
  ],
...
  "paths": {
...
    "/api/v1/buckets/{bucket}/objects/batch": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "post": {
        "tags": [
          "Objects"
        ],
        "operationId": "uploadObjectBatch",
        "summary": "Batch upload files with multipart/form-data",
        "description": "manifest is a JSON string array. Each item declares file_field and relative_path. Additional file parts must use the exact field names referenced by manifest[].file_field.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadBatchForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Files uploaded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBatchEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Multipart payload, manifest, bucket name, prefix, visibility, or final object key is invalid. Error codes include invalid_multipart_request, invalid_batch_manifest, batch_file_missing, invalid_bucket_name, invalid_prefix, invalid_visibility.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token. Error code: unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Bucket not found. Error code: bucket_not_found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the configured upload size. Error code: payload_too_large.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Batch upload failed. Error codes may include bucket_lookup_failed, batch_file_buffer_failed, batch_file_open_failed, object_store_failed, object_metadata_failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "BucketParam": {
        "name": "bucket",
        "in": "path",
        "required": true,
        "description": "Bucket name. Must match ^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$.",
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$",
          "minLength": 3,
          "maxLength": 63
        }
      },
...
    },
...
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "request_id",
          "data"
        ],
        "properties": {
          "request_id": {
            "type": "string",
            "description": "Request correlation ID returned by middleware."
          },
          "data": {
            "type": "object",
            "description": "Endpoint-specific payload."
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "request_id",
          "error"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "Object": {
        "type": "object",
        "required": [
          "id",
          "bucket_name",
          "object_key",
          "original_filename",
          "size",
          "content_type",
          "etag",
          "visibility",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "bucket_name": {
            "type": "string"
          },
          "object_key": {
            "type": "string",
            "maxLength": 512
          },
          "original_filename": {
            "type": "string",
            "maxLength": 255
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "content_type": {
            "type": "string"
          },
          "etag": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UploadBatchForm": {
        "type": "object",
        "required": [
          "manifest"
        ],
        "properties": {
          "prefix": {
            "type": "string",
            "maxLength": 512,
            "description": "Optional folder prefix. Non-empty values must end with '/'."
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ],
            "description": "Visibility applied to every uploaded object. Defaults to private when omitted."
          },
          "manifest": {
            "type": "string",
            "description": "JSON string array of UploadBatchManifestItem objects.",
            "example": "[{\"file_field\":\"file_0\",\"relative_path\":\"index.html\"},{\"file_field\":\"file_1\",\"relative_path\":\"assets/app.js\"}]"
          }
        },
        "additionalProperties": {
          "type": "string",
          "format": "binary",
          "description": "Additional multipart file parts referenced by manifest[].file_field."
        }
      },
      "UploadBatchResult": {
        "type": "object",
        "required": [
          "uploaded_count",
          "items"
        ],
        "properties": {
          "uploaded_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Object"
            }
          }
        }
      },
...
      "UploadBatchEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": [
              "request_id",
              "data"
            ],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/UploadBatchResult"
              }
            }
          }
        ]
      }
    }
  }
}